+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [v0.4.4](https://github.com/0x2142/frigate-notify/releases/tag/v0.4.4) - TBD
+## [v0.4.4](https://github.com/0x2142/frigate-notify/releases/tag/v0.4.4) - Apr 08 2025
  - Add support for notifications via [Matrix](https://frigate-notify.0x2142.com/latest/config/file/#matrix)
  - Add support for `message_thread_id` to [Telegram](https://frigate-notify.0x2142.com/latest/config/file/#telegram) notifications
  - Add review/event link for Pushover action button